+++ trivoliSwimming/config/ventas_con_04.asp
@@ -19,26 +19,40 @@ l_id = request.Form("cabnro")
 Set l_rs = Server.CreateObject("ADODB.RecordSet")
 Set l_cm = Server.CreateObject("ADODB.Command")
 
-'l_sql = "SELECT id"
-'l_sql = l_sql & " FROM calendarios"
-'l_sql  = l_sql  & " WHERE idrecursoreservable = " & l_id
-'rsOpen l_rs, cn, l_sql, 0 
+l_sql = "SELECT id"
+l_sql = l_sql & " FROM cajaMovimientos"
+l_sql  = l_sql  & " WHERE idventaOrigen = " & l_id
+rsOpen l_rs, cn, l_sql, 0 
 
-'if not l_rs.eof then
-'	Response.write "Existen calendarios asociados al M&eacute;dico. No se permite eliminar."
-'	l_rs.close
-'else
+if not l_rs.eof then
+	Response.write "Existen pagos asociados a la venta. No se permite eliminar."
+	l_rs.close
+else
 	l_sql = "DELETE FROM ventas  WHERE id = " & l_id
 
 	l_cm.activeconnection = Cn
 	l_cm.CommandText = l_sql
 	cmExecute l_cm, l_sql, 0
 
+	'Borro detalle de ventas
+	l_sql = "DELETE FROM detalleVentas  WHERE idventa = " & l_id
+
+	l_cm.activeconnection = Cn
+	l_cm.CommandText = l_sql
+	cmExecute l_cm, l_sql, 0
+
+	'Borro costo de ventas
+	l_sql = "DELETE FROM costosVentas  WHERE idventa = " & l_id
+
+	l_cm.activeconnection = Cn
+	l_cm.CommandText = l_sql
+	cmExecute l_cm, l_sql, 0
+	
 	cn.Close
 	Set cn = Nothing
 	
 	Response.write "OK"
-'end if
+end if
